Refactor: Change imports and table name
