+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/user/database/UserEntity.kt
@@ -8,7 +8,7 @@ import javax.persistence.*
 import javax.validation.constraints.NotNull
 
 @Entity
-@Table(name = "user")
+@Table(name = "users")
 class UserEntity(
     @Column(unique = true)
     @NotNull
